requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.14
Django==4.2.2
djangorestframework==3.14.0
frozenlist==1.4.0
future==0.18.3
idna==3.4
line-bot-sdk==3.1.0
multidict==6.0.4
mysqlclient==2.1.1
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
yarl==1.9.2
